+++ detox/src/android/core/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  native: require('./native'),
+  web: {}, // TODO
+};
