+++ detox/src/android/core/native/WaitFor.js
@@ -1,0 +1,73 @@
+const { WaitForInteraction } = require('../../interactions/native')
+const matchers = require('../../matchers/native')
+
+class WaitFor {
+  constructor(invocationManager) {
+    this._invocationManager = invocationManager;
+  }
+}
+
+class WaitForElement extends WaitFor {
+  constructor(invocationManager, element) {
+    super(invocationManager);
+    this._element = element;
+  }
+
+  get not() {
+    this._notCondition = true;
+    return this;
+  }
+
+  toBeVisible() {
+    return new WaitForInteraction(this._invocationManager, this._element, this._notCondition ? new matchers.VisibleMatcher().not : new matchers.VisibleMatcher());
+  }
+
+  toBeNotVisible() {
+    return this.not.toBeVisible();
+  }
+
+  toExist() {
+    return new WaitForInteraction(this._invocationManager, this._element, this._notCondition ? new matchers.ExistsMatcher().not : new matchers.ExistsMatcher());
+  }
+
+  toNotExist() {
+    return this.not.toExist();
+  }
+
+  toHaveText(text) {
+    return new WaitForInteraction(this._invocationManager, this._element, this._notCondition ? new matchers.TextMatcher(text).not : new matchers.TextMatcher(text));
+  }
+
+  toNotHaveText(text) {
+    return this.not.toHaveText(text);
+  }
+
+  toHaveLabel(value) {
+    return new WaitForInteraction(this._invocationManager, this._element, this._notCondition ? new matchers.LabelMatcher(value).not : new matchers.LabelMatcher(value));
+  }
+
+  toNotHaveLabel(value) {
+    return this.not.toHaveLabel(value);
+  }
+
+  toHaveId(value) {
+    return new WaitForInteraction(this._invocationManager, this._element, this._notCondition ? new matchers.IdMatcher(value).not : new matchers.IdMatcher(value));
+  }
+
+  toNotHaveId(value) {
+    return this.not.toHaveId(value);
+  }
+
+  toHaveValue(value) {
+    return new WaitForInteraction(this._invocationManager, this._element, this._notCondition ? new matchers.ValueMatcher(value).not : new matchers.ValueMatcher(value));
+  }
+
+  toNotHaveValue(value) {
+    return this.not.toHaveValue(value);
+  }
+}
+
+module.exports = {
+  WaitFor,
+  WaitForElement,
+};
