+++ detox/src/android/core/native/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./Element'),
+  ...require('./Expect'),
+  ...require('./WaitFor'),
+};
