+++ detox/src/android/core/web/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('./WebElement'),
+  ...require('./WebExpect'),
+};
